fix: keep session projection dependency-light
